\name{readmnist-package}
\alias{readmnist-package}
\alias{readmnist}
\docType{package}
\title{
\packageTitle{readmnist}
}
\description{
\packageDescription{readmnist}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{readmnist}
\packageIndices{readmnist}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{readmnist}

Maintainer: \packageMaintainer{readmnist}
}
\references{
https://stackoverflow.com/questions/21521571/how-to-read-mnist-database-in-r
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
%~~ in the R documentation directory ~~
\keyword{ package }
\examples{
  Info <- Read.mnist(system.file("exdata/t10k-labels.idx1-ubyte", package="readmnist"))
}

